backend-files/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
torch==2.6.0
torchvision==0.21.0
pillow==10.1.0
numpy==1.26.2
python-dotenv==1.0.0
kagglehub==0.2.7
scikit-learn==1.3.2 